--- pick-front/src/views/MainView.vue
@@ -1,14 +1,0 @@
-<template>
-    <div class="main-view">
-        <h1>메인 페이지 입니다 🏠</h1>
-        <MatchingSearch/>
-    </div>
-</template>
-
-<script setup>
-    import MatchingSearch from '@/matching/components/MatchingSearch.vue';
-</script>
-
-<style scoped>
-
-</style>
